Limitar como maximo a 4 ataques a los jugadores y puede tener menos cantidad de ataques de lo establecido, mas cambiar los nombres
