--- DemonLoopCode/Assets/Scripts/3D/Enemy_Mechanics.cs
@@ -1,115 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Enemy_Mechanics : MonoBehaviour
-{
-    [Header("View conditions")]
-    [SerializeField] private LayerMask layer;
-    [SerializeField] private float distance = 100f;
-
-    [Header("Enemy patrol")]
-    [SerializeField] bool patrol = false;
-    [SerializeField] float speed = 10f;
-    [SerializeField] GameObject puntoA;
-    [SerializeField] GameObject puntoB;
-    
-
-    bool change = false;
-    bool attack = false;
-
-    Vector3 direction = -Vector3.forward;
-
-    Enter_Battle _Battle;
-
-    void Start()
-    {
-        if (patrol)
-        {
-            StartCoroutine(Enemy_patrol());
-        }  
-    }
-
-    // Update is called once per frame
-    void FixedUpdate()
-    {
-        // En el caso que no haya guardado el script Lo busca  hasta encontrarlo (Esto se debe aque si por error coge el 'System' el cual va a ser destruido este se volvera de nuevo null y si se hace una vez no ira a por el nuevo 'System').
-        if (_Battle == null)
-        {
-            _Battle = GameObject.Find("System").GetComponent<Enter_Battle>();
-        }
-
-        RaycastHit hit;
-
-        // Crea una linea invisible de una determinada distancia el cual actua como un trigger y solo detecta los objetos que se encuentran en las capas selecionadas.
-        if (Physics.Raycast(transform.position, direction, out hit, distance, layer))
-        {
-            Debug.DrawLine(transform.position, hit.transform.position, Color.green);
-
-            attack = true;
-
-            hit.transform.GetComponent<Player_Move>().enabled = false; // Desactiva el movimiento del jugador.
-
-            // Si encuentra alguno de los dos componentes en el objeto jugador los desactiva antes de iniciar la pelea.
-            if (hit.transform.GetComponent<KeyBoard_Controls>())
-                hit.transform.GetComponent<KeyBoard_Controls>().enabled = false;
-            if (hit.transform.GetComponent<Controller_Controls>())
-                hit.transform.GetComponent<Controller_Controls>().enabled = false;
-
-
-            // Antes de activar la pelea el enemigo se acerca al jugador.
-            if (Vector3.Distance(transform.position, hit.transform.position) < 1.5f) // Inicia el combate.
-            {
-                _Battle.StartBattle(transform.parent.gameObject/*, false*/);
-            }
-            else
-            {
-                transform.position = Vector3.MoveTowards(transform.position, hit.transform.position, speed * Time.deltaTime);
-            }
-        }
-        else
-        { Debug.DrawRay(transform.position, direction * distance, Color.red); }
-    }
-
-    IEnumerator Enemy_patrol()
-    {
-        while (true)
-        {
-            if (!attack) // Cuando detecta al jugador este deja de estar en modo 
-            {
-                if (change)
-                {
-                    transform.position = Vector3.MoveTowards(transform.position, puntoB.transform.position, speed * Time.deltaTime);
-                }
-                else
-                {
-                    transform.position = Vector3.MoveTowards(transform.position, puntoA.transform.position, speed * Time.deltaTime);
-                }
-
-                if (Vector3.Distance(transform.position, puntoA.transform.position) < 0.0001f)
-                {
-                    transform.rotation = Quaternion.Euler(0, 90, 0);
-                    direction = -Vector3.forward;
-                    change = true;
-                }
-
-                if (Vector3.Distance(transform.position, puntoB.transform.position) < 0.0001f)
-                {
-                    transform.rotation = Quaternion.Euler(0, -90, 0);
-                    direction = Vector3.forward;
-                    change = false;
-                }
-            }
-
-            yield return new WaitForSeconds(0.00001f);
-        }
-    }
-
-    private void OnTriggerEnter(Collider other)
-    {
-        if (other.transform.name == "Player")
-        {
-            _Battle.StartBattle(transform.parent.gameObject/*, true*/);
-        }
-    }
-}
